+++ bcmc/src/main/res/layout/bcmc_view.xml
@@ -62,7 +62,7 @@
         android:layout_height="wrap_content"
         android:layout_weight="1">
 
-        <com.adyen.checkout.card.internal.ui.view.ExpiryDateInput
+        <com.adyen.checkout.ui.core.internal.ui.view.ExpiryDateInput
             android:id="@+id/editText_expiryDate"
             style="@style/AdyenCheckout.Card.ExpiryDateInput"
             android:nextFocusRight="@+id/editText_securityCode"
