+++ card/api/card.api
@@ -564,9 +564,6 @@ public final class com/adyen/checkout/card/internal/provider/CardComponentProvid
 public final class com/adyen/checkout/card/internal/ui/DefaultCardDelegate$Companion {
 }
 
-public final class com/adyen/checkout/card/internal/ui/model/ExpiryDate$Companion {
-}
-
 public final class com/adyen/checkout/card/internal/ui/model/InstallmentOptionParams$CardBasedInstallmentOptions : com/adyen/checkout/card/internal/ui/model/InstallmentOptionParams {
 	public fun <init> (Ljava/util/List;ZLcom/adyen/checkout/card/CardBrand;)V
 	public final fun component1 ()Ljava/util/List;
@@ -610,21 +607,6 @@ public final class com/adyen/checkout/card/internal/ui/view/CardNumberInput$Comp
 }
 
 public final class com/adyen/checkout/card/internal/ui/view/CardView$Companion {
-}
-
-public final class com/adyen/checkout/card/internal/ui/view/ExpiryDateInput : com/adyen/checkout/ui/core/internal/ui/view/AdyenTextInputEditText {
-	public static final field Companion Lcom/adyen/checkout/card/internal/ui/view/ExpiryDateInput$Companion;
-	public static final field SEPARATOR Ljava/lang/String;
-	public fun <init> (Landroid/content/Context;)V
-	public fun <init> (Landroid/content/Context;Landroid/util/AttributeSet;)V
-	public fun <init> (Landroid/content/Context;Landroid/util/AttributeSet;I)V
-	public synthetic fun <init> (Landroid/content/Context;Landroid/util/AttributeSet;IILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public fun afterTextChanged (Landroid/text/Editable;)V
-	public final fun getDate ()Lcom/adyen/checkout/card/internal/ui/model/ExpiryDate;
-	public final fun setDate (Lcom/adyen/checkout/card/internal/ui/model/ExpiryDate;)V
-}
-
-public final class com/adyen/checkout/card/internal/ui/view/ExpiryDateInput$Companion {
 }
 
 public final class com/adyen/checkout/card/internal/ui/view/SecurityCodeInput : com/adyen/checkout/card/internal/ui/view/CardNumberInput {
