+++ card/src/main/res/layout/card_view.xml
@@ -100,7 +100,7 @@
             android:layout_marginEnd="@dimen/standard_half_margin"
             android:layout_weight="1">
 
-            <com.adyen.checkout.card.internal.ui.view.ExpiryDateInput
+            <com.adyen.checkout.ui.core.internal.ui.view.ExpiryDateInput
                 android:id="@+id/editText_expiryDate"
                 style="@style/AdyenCheckout.Card.ExpiryDateInput"
                 android:nextFocusDown="@id/editText_securityCode"
