+++ ui-core/api/ui-core.api
@@ -217,6 +217,9 @@ public final class com/adyen/checkout/ui/core/internal/ui/model/AddressParams$Po
 	public fun toString ()Ljava/lang/String;
 }
 
+public final class com/adyen/checkout/ui/core/internal/ui/model/ExpiryDate$Companion {
+}
+
 public final class com/adyen/checkout/ui/core/internal/ui/view/AddressFormInput : android/widget/LinearLayout {
 	public fun <init> (Landroid/content/Context;)V
 	public fun <init> (Landroid/content/Context;Landroid/util/AttributeSet;)V
@@ -269,6 +272,21 @@ public final class com/adyen/checkout/ui/core/internal/ui/view/AdyenTextInputEdi
 
 public abstract interface class com/adyen/checkout/ui/core/internal/ui/view/AdyenTextInputEditText$Listener {
 	public abstract fun onTextChanged (Landroid/text/Editable;)V
+}
+
+public final class com/adyen/checkout/ui/core/internal/ui/view/ExpiryDateInput : com/adyen/checkout/ui/core/internal/ui/view/AdyenTextInputEditText {
+	public static final field Companion Lcom/adyen/checkout/ui/core/internal/ui/view/ExpiryDateInput$Companion;
+	public static final field SEPARATOR Ljava/lang/String;
+	public fun <init> (Landroid/content/Context;)V
+	public fun <init> (Landroid/content/Context;Landroid/util/AttributeSet;)V
+	public fun <init> (Landroid/content/Context;Landroid/util/AttributeSet;I)V
+	public synthetic fun <init> (Landroid/content/Context;Landroid/util/AttributeSet;IILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public fun afterTextChanged (Landroid/text/Editable;)V
+	public final fun getDate ()Lcom/adyen/checkout/ui/core/internal/ui/model/ExpiryDate;
+	public final fun setDate (Lcom/adyen/checkout/ui/core/internal/ui/model/ExpiryDate;)V
+}
+
+public final class com/adyen/checkout/ui/core/internal/ui/view/ExpiryDateInput$Companion {
 }
 
 public final class com/adyen/checkout/ui/core/internal/ui/view/PaymentInProgressView : androidx/constraintlayout/widget/ConstraintLayout, com/adyen/checkout/ui/core/internal/ui/ComponentView {
