+++ ui-core/src/main/java/com/adyen/checkout/ui/core/internal/ui/model/ExpiryDate.kt
@@ -1,11 +1,11 @@
 /*
- * Copyright (c) 2019 Adyen N.V.
+ * Copyright (c) 2024 Adyen N.V.
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by arman on 16/9/2019.
+ * Created by ararat on 17/7/2024.
  */
-package com.adyen.checkout.card.internal.ui.model
+package com.adyen.checkout.ui.core.internal.ui.model
 
 import androidx.annotation.RestrictTo
